Pull out another component into separate file
